+++ src/Burger.js
@@ -1,0 +1,10 @@
+function Burger() {
+    return (
+      <div className="nav__burger burger">
+        <span className="burger__line" />
+        <span className="burger__line" />
+        <span className="burger__line" />
+      </div>
+    );
+  }
+  export default Burger;
