+++ src/NavMenu.js
@@ -1,0 +1,25 @@
+import "./css/style.css";
+function NavMenu() {
+    return (
+      <div className="nav__menu menu">
+        <ul className="menu__list">
+          <li className="menu__item">
+            <a href="http://" className="menu__link">
+              Главное
+            </a>
+          </li>
+          <li className="menu__item">
+            <a href="http://" className="menu__link">
+              Мой плейлист
+            </a>
+          </li>
+          <li className="menu__item">
+            <a href="http://" className="menu__link">
+              Войти
+            </a>
+          </li>
+        </ul>
+      </div>
+    );
+  }
+export default NavMenu;
