+++ src/PlaylistTitle.js
@@ -1,0 +1,16 @@
+function PlaylistTitle() {
+    return (
+      <div className="content__title playlist-title">
+        <div className="playlist-title__col col01">Трек</div>
+        <div className="playlist-title__col col02">ИСПОЛНИТЕЛЬ</div>
+        <div className="playlist-title__col col03">АЛЬБОМ</div>
+        <div className="playlist-title__col col04">
+          <svg className="playlist-title__svg" alt="time">
+            <use xlinkHref="img/icon/sprite.svg#icon-watch" />
+          </svg>
+        </div>
+      </div>
+    );
+  }
+  export default PlaylistTitle;
+
