+++ src/example stratos font/__MACOSX/._example stratos font
