+++ src/example stratos font/__MACOSX/example stratos font/._Stratos-Bold.woff
